Fix the new wptmetadata-bot for manifest updates (#4115)
Problem 1: New autoapprove job cannot infer the PR that started it

The new bot ran into some problems upon initial rollout. It is meant to
use the principle of least privilege. This helpful for non-verified
actions. More about it here: https://github.com/peter-evans/create-pull-request/blob/main/docs/concepts-guidelines.md#push-pull-request-branches-to-a-fork

During the rollout, it was discovered that the new workflow_run actually
does not have access to which pull request it came from. (There will
initial attempts to read from the event payload but it failed).

GitHub itself documents how to get the PR information to the subsequent
workflows:
- https://securitylab.github.com/research/github-actions-preventing-pwn-requests/

This commit implements that solutions.
Implementation based on what GitHub suggested:
- Save the PR number in test.yml
- use the PR number in auto-approve.yml

Problem 2: automerge workflow does not work with forked contributions

Forks do not have access to repository secrets during the following
events (which the automerge workflow uses):
- pull_request
- pull_request_review

As a result, the automerge workflow will not work on this. We get
around this by now using the now built in GitHub auto merge feature.
By enabling it, it does not enable auto merge for PRs. Instead, it
presents the option that developers to enable automerge. If enabled,
PRs will auto merge once the branch protection rules are fulfilled.

Implementation:
- Modified the update_wpt_manifest with a separate job which will enable auto merge once it is created.
- Added a note to the pull request template about human developers not
  using the automerge feature.

---

Requirements after this is merged:
- Enable the option for contributors to enable auto-merge
  - Human developers should not use it and there has been a note added
    to the PR template.
- Recreate the fork for wptmetadata-bot
- Keep the branch requirement that a reviewer must approve and that the
  "test" action must pass