Remove dedicated-worker-import-data-url.any(.worker).html and shared-worker-import-data-url.window.html from interop-2023-modules
Fixing https://github.com/web-platform-tests/interop/issues/406

In this test, The worker will load a data URI, which import another test
script statically, and the imported script will import 'export-on-load-script.js',
which will be served with the CORS header 'Access-Control-Allow-Origin: *' in
/workers/modules/resources/export-on-load-script.js.headers

In the test case [Static import (redirect).], 'export-on-load-script.js' will be
redirected, and the redirected script doesn't have the CORS header.

https://github.com/web-platform-tests/wpt/blob/2cd449b9df6f92862abac0143ac6fe674fd5278e/workers/modules/resources/import-test-cases.js#L13
https://github.com/web-platform-tests/wpt/blob/2cd449b9df6f92862abac0143ac6fe674fd5278e/workers/modules/resources/static-import-redirect-worker.js#L3

The script "export-on-load-script.js" is same-origin with the
document, but not same-origin with the worker script which is loaded by
data: URI.

Due to the spec issue, https://github.com/whatwg/html/issues/9571
There isn't an agreement between browser vendors on which [settings object](https://html.spec.whatwg.org/multipage/webappapis.html#environment-settings-object)
to use when fetching the descendants of module worker.

https://wpt.fyi/results/workers/modules/dedicated-worker-import-data-url.any.html?label=experimental&label=master&aligned
https://wpt.fyi/results/workers/modules/dedicated-worker-import-data-url.any.worker.html?label=experimental&label=master&aligned
https://wpt.fyi/results/workers/modules/shared-worker-import-data-url.window.html?label=experimental&label=master&aligned

So removing these three tests from interop-2023-modules.
(According to the README.md in wpt-metadata, when a 'label' is used,
'subtest' is omitted, so I remove these three files from
interop-2023-modules completely)