Update Interop metadata for test splits
In https://github.com/web-platform-tests/wpt/pull/41985 we are proposing
to rename and split some tests, in order to increase test coverage for
URL in ShadowRealm scopes. ShadowRealm is irrelevant to Interop, but we
do have to split some files that previously contained a mix of tests
both suitable and unsuitable for executing in ShadowRealm scopes.

This PR adjusts the `interop-2023-url` label so that the splits in the
abovementioned PR have as little effect as possible on the Interop
scores.

It also adjusts the browser bug metadata so that the tests referred to
are in the correct files.